requirements.txt
torch
torchvision
transformers
shapiq
pillow
requests
scikit-image
jupyterlab      # oder notebook
